+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 
+## 0.1.1
+
+ - Added binutils 2.27, expat 2.2.0, mesa 13.0.0, and llvm 4.0.0 (in-development)
+ - Improved mason.cmake to support packages that report multiple static libs
+ - Improved llvm >= 3.8.1 packages to support `-flto` on linux
+
 ## 0.1.0
  - First versioned release